feat: add more composition
