+++ lib/util.js
@@ -25,10 +25,14 @@ const getSortedQS = (obj) => {
   return qs.stringify(_obj);
 };
 
+const compose2 = (f, g) => (...args) => f(g(...args));
+const pipe = (...fns) => fns.reduceRight(compose2);
+
 module.exports = {
   hexToBuf,
   bufToHex,
   dumpBuf,
   randomBytes,
-  getSortedQS
+  getSortedQS,
+  pipe
 };
